# Faster propertynames when no dynamic properties are present