--- a/src/dynamic.jl
+++ b/src/dynamic.jl
@@ -118,16 +118,16 @@ macro dynamic(expr::Expr)
 
     return quote
         $(esc(expr))
-        function Base.propertynames(x::$struct_name)
-            if isempty($property_dict(x)) 
+        function Base.propertynames(x::$(esc(struct_name)))
+            if isempty(property_dict(x)) 
                 fieldnames(typeof(x))[1:end-1]
             else
-                (fieldnames(typeof(x))[1:end-1]..., keys($property_dict(x))...)
+                (fieldnames(typeof(x))[1:end-1]..., keys(property_dict(x))...)
             end
         end
-        function Base.propertynames(x::$struct_name, private::Bool)
+        function Base.propertynames(x::$(esc(struct_name)), private::Bool)
             if private
-                (fieldnames(typeof(x))..., keys($property_dict(x))...)
+                (fieldnames(typeof(x))..., keys(property_dict(x))...)
             else
                 Base.propertynames(x)
             end