@AntonOresten 2024-08-25
Neat! I hadn't thought of this.

Well done, and thanks for contributing!