@Tortar 2024-08-25
thanks for the package, as I said I think it can be very useful, actually I'm trying to push for its integration in Agents.jl: https://github.com/JuliaDynamics/Agents.jl/issues/1070